upd: Lesson 4C - Loaders finished. Recap
	What I learned

This section is much more easier than the previous one.

First of all, we changed our ArrayAdapter to CursorAdapter, that populates view from Cursor, that is more convenient when interacting with ContentProviders. We also added Utility class to provide some helpful functions, and store them seperately, I think it goes to "best practice".

And the main part, is using CursorLoader. The steps are not hard:
1. Define Constant ID for Loader
2. Make class implement LoaderCallback<with "Cursor" here> and implement 3 main methods(those are also not hard):
	* main - onCreateLoader - just return CursorLoader with the URI we need to fetch and columns(for efficiency) we need.
	* onLoadFinished - here, we just make our adapter change data with swapCursor method
	* onLoaderReset - just make swapCursor take parameter null, saying to reset it
3. in onActivityCreated method, just call getLoaderManager().initLoader(LOADER_ID, null, this) - and here we are

Main points to point OUT here:
* no content resolver, content provider and database done on UI
* CursorLoader handles everything in Background thread, we just give URI and wait while it does all job

So, CursorLoader goes to "Best Practice".