feat: Finish Layout designing with wireframes
		What is done

Now we have responsive design. In tablets we have Two Panes, in Smart Phones it is One Pane. For landscape, as well as for Tablets we have different elements places. The way it is done :
* using different folders for different screen sizes(layout-sw600dp, layout-land, values-sw600dp, etc):
		* sw600dp - stands for Smalles Width 600dp
* using references

Interface implemented, when choosing Item in ListView. MainActivity implements Callback from ListView. It is called when List item clicked. Overriden in MainActivity, it makes different actions based on Two or One pane. For Two Pane it reokaces Detail Fragment, for One Pane it start Intent to DetailActivity.

Detail Fragment now has incoming Bundle where Uri to data is stored. We update Loader after this.