feat: Add Services
This is first commit in Lesson 6. The commit is about Services. Instead of FetchWeather AsyncTask, new Service added, that does the same thing. We prefer Service, because AsyncTask is tied with UI, and everytime onCreate called(e.g. on screen rotation), new instance of AsyncTask is created, and it does the same thing. So, to get rid of this, we created SunshineService, and put all FetchWeather task there. Our Service is IntentService. You just override onHandleIntent method, and that's it. Call it like Activity, but startService(intent).

To read more about Services, visit http://developer.android.com