feat: Add SyncAdapter
Before, I though that SyncAdapters are hard to understand. But throughout this course I understand that it is not that hard. But first, what for are SyncAdapters, here is the definition from course:

SyncAdapters are meant to keep local data on the device in sync with data on the web.

With SyncAdapter, we are able to Sync our local data with data in Cloud. We also can take advantage of using Accounts, but sometimes there is no need to authenticate when fetching data. But in SyncAdapter implementation we need Accounts. We just create dummy Account and put there stub code. Here are steps:
1. Create dummy Authenticator
2. Create Service for that Authenticator
3. Create Service for SyncAdapter
4. And SyncAdapter itself, where in onPerformSync() method the main staff is done. To start SyncAdapter just call ContentResolver.requestSync(bla bla bla).

Of course, there are some staff to write in AndroidManifest, as registering Services. And also in xml we should add our authenticator and sync-adapter.

Currently, our SyncAdapter is called in method updateWeather(), but further we will develop it in a better way.