feat: Update our SyncAdapter, so that is Syncs Periodically
Here is the "best practice". We do Syncronization periodically, to ensure UX. How we did it:

We configure periodic sync with SyncRequest class. And give it as parameter to requestSync method. Then setSyncAutomatically. And start syncing. After this, Sync will go on its own periodically.

All staff is in the code.